requirements.txt
discord.py>=2.0
asyncio
json
pytest
python-dotenv
```